+++ src/plugins/discovery/discovery/discovery.error.ts
@@ -1,4 +1,4 @@
-import { NFError } from "../../lib/native-federation.error";
+import { NFError } from "../../../lib/native-federation.error";
 
 class NFDiscoveryError extends NFError {
     constructor(message: string) {
