+++ src/plugins/discovery/index.ts
@@ -3,5 +3,5 @@ export { initFederationFromDiscovery } from './init-federation-from-discovery';
 /**
  * Mappers
  */
-export * from './noop.mapper';
-export * from './manifest/manifest.mapper';
+export * from './mapper/noop.mapper';
+export * from './mapper/manifest.mapper';
