+++ src/plugins/discovery/mapper/noop.mapper.ts
@@ -1,4 +1,4 @@
-import type { DiscoveredRemotes, DiscoveryMapper } from "./discovery.contract";
+import type { DiscoveredRemotes, DiscoveryMapper } from "../discovery/discovery.contract";
 
 
 type NoopMapper = DiscoveryMapper<DiscoveredRemotes, DiscoveredRemotes>;
